requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
